Added Editing Capability
The editing works mostly now. The design decision is not the best
possible and hopefully we will be able to improve it. It relies on
passing an index around between view controllers. Additionally, because
of the way it’s implemented, closing the edit form takes you back to
the root view controller instead of just going back to the view detail
controller, which would probably make more sense, because I couldn’t
figure out how to refresh the information in the detail view. Hopefully
will find a solution to this.